finalizado 95

+++ tapeteyoga/src/main/java/br/senac/ecommerce/tapeteyoga/model/Frete.java
@@ -7,12 +7,12 @@ import lombok.Data;
 @Data
 public class Frete {
 
-    private Long id;
+    private BigDecimal id;
     private String tipoEntrega;
     private BigDecimal valor;
     private Integer dias;
 
-    public Frete(Long id, String tipoEntrega, BigDecimal valor, Integer dias) {
+    public Frete(BigDecimal id, String tipoEntrega, BigDecimal valor, Integer dias) {
         this.id = id;
         this.tipoEntrega = tipoEntrega;
         this.valor = valor;
